refactor: prepare crypto extension
